**Summary:** Similar to https://github.com/pytorch/ao/pull/2986,
this commit improves the prepare vs convert SQNR of NVFP4 QAT
from 12 to inf. This is achieved by refactoring NVFP4 QAT to
mimick the PTQ numerics exactly, using a new linear class to
incorporate both the quantization and mm logic.

**Test Plan:**
```
python test/quantization/test_qat.py -k test_qat_nvfp4
python test/quantization/test_qat.py -k test_quantize_api_nvfp4
```

End-to-end tests TBD.

[ghstack-poisoned]

+++ torchao/prototype/mx_formats/nvfp4_tensor.py
@@ -815,4 +815,4 @@ def nvfp4_quantize(
     # TODO: NotImplementedError: "copy_kernel" not implemented for 'Float4_e2m1fn_x2'
     # data_lp = pack_uint4(data_lp).view(torch.float4_e2m1fn_x2)
     data_lp = pack_uint4(data_lp)
-    return out_scales.to(torch.float8_e4m3fn), data_lp
+    return out_scales, data_lp
